feat: support multiple execution environments
