+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@limechain/rollup-codes-mcp",
-  "version": "1.0.3",
+  "version": "1.0.4",
   "main": "build/index.js",
   "type": "module",
   "bin": "build/index.js",
